allow setting the initial seed in the config
