+++ Cargo.toml
@@ -37,6 +37,7 @@ ringlog = "0.1.1"
 serde = "1.0.144"
 serde_json = "1.0.94"
 session = { git = "https://github.com/pelikan-io/pelikan" }
+sha2 = "0.10.7"
 slab = "0.4.8"
 tokio = { version = "1.21.2", features = ["full"] }
 tokio-boring = "2.1.5"
